requirements.txt
numpy==1.24.3
opencv_python_headless==4.8.0.74
Pillow==10.3.0
streamlit==1.34.0
streamlit_theme==0.58.0
torch==2.2.0
ultralytics==8.2.8
